--- src/data/data.js
@@ -1,12 +1,0 @@
-var Data=[
-  {
-  id: 0,
-    name: '第四届微观摄影大赛',
-      publisher: '植物保护学院',
-        startTime: '2019-3-11 8:00:00',
-          endTime: '2019-3-22 8:00:00-22:00:00'
-  },
-]
-module.exports = {
-  postlist: Data
-}
