+++ src/pages/index/index.json
@@ -1,9 +1,1 @@
-      {
-      "window": {
-        "navigationBarTitleText": "报名表",
-            "navigationBarBackgroundColor": "#ffffff",
-            "navigationBarTextStyle": "black",
-            "backgroundColor": "#eeeeee",
-            "backgroundTextStyle": "light"
-      }
-      }
+{}
