--- src/pages/index/index.wxml
@@ -1,33 +1,0 @@
-<!--index.wxml-->
-<view>
-<view class='container'>
-  <form bindsubmit="back_houtai">
-    <view class='row'>
-      <view class='left'>姓名</view>
-      <input type='text' name='userName' placeholder='请输入姓名' value='{{userName}}'/>
-    </view>
-    <view class='row'>
-      <view>性别 </view>
-     <radio-group>
-     <label><radio value='0' >女</radio></label>
-     <label><radio value='1'>男</radio></label>
-     </radio-group>
-    </view>
-    <view class='row'>
-      <text>学号 </text>
-      <input type='text' name='stuid' placeholder='请输入学号' value='{{stuid}}'/>
-    </view>
-    <view class='row'>
-      <text>身份证号 </text>
-      <input type='text' name='idcard' placeholder='请输入身份证号' value='{{idcard}}'/>
-    </view>
-       <view class='row'>
-      <text>手机号 </text>
-      <input type='text' name='phone' placeholder='请输入手机号' value='{{phone}}'/>
-    </view>
-    <view>
-      <button form-type='submit' type='primary'>提交</button>
-    </view>
-  </form>
-</view>
-</view>
