--- src/pages/index/index.wxss
@@ -1,35 +1,0 @@
-container{
-
-    padding: 1rem;
-    font-size: 0.8rem;
-    line-height: 2rem;
-    border-shadow: 1px 1px #0099CC;
-}
-.row{
-    display: flex;
-    align-items: center;
-    margin-bottom: 0.8rem;
-}
-.row text{
-  flex-grow: 0;
-  text-align: right;
-}
-.row input{
-    font-size: 0.7rem;
-    color: #ccc;
-    flex-grow: 3;
-    border: 1px solid #0099CC;
-    display: inline-block;
-    border-radius: 0.3rem;
-    box-shadow: 0 0 0.15rem #aaa;
-    padding: 0.3rem;
-}
-/*placeholder() {
-      display: block;
-      width: 100%;
-      height: 180rpx;
-      overflow: hidden;
-    }*/
-.row button{
-    padding: 0 2rem;
-}
