+++ src/pages/logs/index.wxml
@@ -1,4 +1,4 @@
-<navigator url="../index/index?id=1" view wx:for="{{list}}" wx:key="list">
-<view class='title_style'>活动：{{list.name}}</view>
-<view class='date_style'>主办方：{{list.publisher}}</view>
+<navigator url="../logs/logs?id=1" view wx:for="{{list}}" wx:key="list">
+<view class='title_style'>活动：{{item.name}}</view>
+<view class='date_style'>主办方：{{item.publisher}}</view>
 </navigator>
