+++ src/pages/logs/logs.json
@@ -1,4 +1,9 @@
-{
-  "navigationBarTitleText": "查看启动日志",
-  "usingComponents": {}
-}
+      {
+        "window": {
+          "navigationBarTitleText": "报名表",
+          "navigationBarBackgroundColor": "#ffffff",
+          "navigationBarTextStyle": "black",
+          "backgroundColor": "#eeeeee",
+          "backgroundTextStyle": "light"
+        }
+      }
